EJOIN: Support nodeset filtering for node consolidation/matching
