+++ packages/seacas/applications/ejoin/EJ_mapping.h
@@ -1,4 +1,4 @@
-// Copyright(C) 1999-2020, 2022 National Technology & Engineering Solutions
+// Copyright(C) 1999-2020, 2022, 2025 National Technology & Engineering Solutions
 // of Sandia, LLC (NTESS).  Under the terms of Contract DE-NA0003525 with
 // NTESS, the U.S. Government retains certain rights in this software.
 //
@@ -21,3 +21,7 @@ void build_local_element_map(RegionVector &part_mesh, std::vector<INT> &local_el
 template <typename INT>
 void generate_element_ids(RegionVector &part_mesh, const std::vector<INT> &local_element_map,
                           std::vector<INT> &global_element_map);
+
+template <typename INT>
+void select_nodeset_nodes(RegionVector &part_mesh, std::vector<INT> &local_node_map,
+                          const Omissions &nset_match);
